The early bird can be lazy later... (Refactor metadata and add initial support for proposed compiled form)
This change creates read-only interfaces for metadata which can be implemented in different ways for different purposes. One proposed implementation is a compiled form of metadata that can be created by Rosyln when the application assembly is built. A hand-coded version of this is checked in to the functionals with infrastructure types added to the product to support this. This is used to verify the pay-per-play memory use of the compiled metadata.

Also included here are:
- Some initial moves in the direction of shadow state
- No longer caching PropertyInfos
for handling other types, but we should wait to see if they are needed
- After discussion Entity is renamed to EntityType
- Model no longer has Name/StorageName
- Annotations are restricted to strings for now. We have ideas