Are you the Keymaster? (First attempt at wiring up change tracker to identity generation)
The idea here is to allow the change tracker to generate a key for new entities when added. Whether or not a key should be generated is determined by metadata on the property in the model. If it is known that key generation is required then the provider must be involved in selecting the appropriate key generator for the type of the property. Providers register an IdentityGeneratorFactory to handle this. However, the user can also register an IdentityGeneratorFactory to override the behavior for some or all properties of a given type.

We could also allow for an identity generator to be registered directly in the model, but this has implications for serialization of the model which is needed for both Migrations and compiled model state. It may be that there is some hook that allows runtime annotation of the model even if it has been compiled.

The configuration keeps the collection of active identity generators so that they can be shared by multiple context instances. This can also be changed in the configuration to allow for different levels of sharing.

Some additional testing still required.