30 is the new 36 (Refactoring of compiled metadata)
This change again uses ideas from Emil to further reduce heap memory usage in the compiled model. The main idea is to only store references to the arrays themselves to the avoid the overhead of a heap-allocated wrapper object. Note that it would be possible to not store a reference for the keys but to build the enumerable each time it is needed, but we decided not to do this at least for now.

Additional benefits include fewer and easier to read/generate generated classes. This helps with the assembly size because there is assembly overhead for each type used.