Gaining some vertical height... (Allowing context to create a model based on the declared EntitySet properties)
This gets a bit more of the vertical slice needed for MusicStore in place by allowing the context to build a model based on its EntitySet properties. There is also a temporary convention for bringing in properties and keys of the entities. Some stuff here will change, but it gets some things we need working so we can get closer to an end-to-end scenario.