Please welcome Mr. Morden... (Supporting shadow state in the change tracker)
We primarily expect that an object model will be used to define and interact with EF. However, the plan for flexible mapping (such as many-to-many and hidden FKs) requires that properties or entire entities do not map directly to CLR properties or types. Instead the state of the properties for these entities existes in a seperaate buffer that travels with the state entry--sometimes known as shadow state. This set of changes introduces basic support for shadow state in the change tracker. Some additional testing still required here.