Talking back to the principal... (Update fixup-related code to use previous changes)
Previous checkins included back-pointers on FKs and key types. This change cleans up the code for finding principals and dependents and using them in fixup such that it makes use of these facilities.

Also included here is the concept of RuntimeModel which implements IModel but adds additional services that require state but which are not part of the model definition itself.