Need for (RS)-1-phenylpropan-2-amine... (Make creation of context instances fast)
EntitySet properties on the context are automatically initialized to appropriate EntitySet instances by the EntityContext constructor. It is important that this is very fast so that the overhead of creating a context instance is small and people will feel okay about using short-lived contexts. Also, I did this work in order to have some concrete idea of the perf impact of using a new configuration per context instance. For this case it is significant--approximately 50 times faster with caching on the configuration.