Expressive fielding... (Add materialization of state entries and entities)
This change allows the StateManager to create a new entry from a value buffer which may contain some shadow properties. The state entry can then materialize an entity from the CLR values in the buffer leaving only shadow values in a new smaller buffer. Creation of the entity instance can come from the compiled model or from a delegate compiled from an expression tree and cached.