It's a high-quality Amstrad... (Updates to EntityConfiguration and top-level sugar)
This change includes introduction of a builder for EntityConfiguration and a change in the way EntityConfiguration works so that it directly modifies the service collection. This allows a service that has been changed to still be injected correctly into other services that depend on it.

Using this there is now a default cache for configurations keyed by context type and an OnConfiguring method where provider-specific or other kinds of functionality can be added.