My manager delegates navigation... (Update navigations to use similar CLR access patterns as elsewhere)
This means that we have cached delegate-based access (without code generation) by default with the option to override in the compiled model. Also, the pattern allows more general access than just ICollections, although this is not yet supported for delegate-based access.

Also, fixup now uses this mechanism and avoids adding duplicates to collections.