Minty-fresh context services... (Create DI scope for context and services it uses, such as StateManager)
This change attempts to address several things:
- StateManager was getting big and involved in many responsibilities that it didn't really need to be involved with. I wanted to simplify the StateManager down to less stuff.
- One of the reasons that the StateManager was getting all this stuff was that it was used by state entries, and state entries don't want to carry around a lot of direct references to services because they are memory sensitive.
- The StateManager is scoped by context instance, and this was handled by using a factory. However, this prevents other services from depending on StateManager, becaus its not in the container, or the factory, because they would get a new instance, not the context one.

The proposed solution is to create a new configuration type which is created in context scope. This configuration derives from EntityConfiguration for global services and then adds sugar for scoped services. This provides the single object that state entries can hold onto and get whatever they need without relying on the StateManager. Also, it allows StateManager and other services to be created in this scope so that they can be properly injected into services that need them and the DI container will handle correct scoping.