End-to-end EntityContext/InMemoryDataStore test
Adding a test to make sure an InMemoryDataStore instance can be shared
between context instances.
This was not working recently when we were not caching the model created
in OnModelCreating because InMemoryDataStore keys of instance of
IEntityType which was a different instance of the same info between
context instances. Adding a test to make sure we don't break it again.