Avoid deleting unchanged entities
RelationalDataStore didn't filter out unchanged entries and the default
operation for ModificationOperation was DELETE. This meant we would
delete any unchanged data from the database.

Made BatchExecutor an explicit dependency of RelationalDataStore so that
it could be mocked for unit testing. Had to make a few classes public to
do this, but discussed with Pawel and the plan was that these would be
public in the end anyway.