We're watching you... (Add support for change detection)
This checkin adds support for original value tracking and change detection to state entries.

We need original values for two purposes:
- Use in the update pipeline for concurrency tokens and graph ordering
- Detecting changes when using snapshot change tracking

Also, original values can be stored eagerly (snapshotting when then entity is materialized) or lazily (only when an attempt is made to change the current value).

We attempt to not store original values when not needed and to store them lazily when possible. This is fully possible for entities that implement both INotifyPropertyChanging and INotifyPropertyChanged.

If an entity implements only INotifyPropertyChanged but not INotifyPropertyChanging then only original values for updates are tracked by default but they are tracked eagerly because we don't know that a change is being made until it has already been made.

If an entity doesn't implement INotifyPropertyChanged, then original values for all properties are tracked eagerly so that DetectChanges can compare current and original values.

A flag on the entity type allows eager tracking of all original values for those users who always want original values in order to implement things like rejecting changes to entities.