...and I am a material...err...maybe not. (Allow strongly-typed non-boxing materialization)
This change introduces an abstraction between the data store and the materializer such that the generic non-boxing ReadField method can be used my the materializer to extract values without boxing. However, the abstraction can also be used to read all values using GetValues. This provides flexibility such that if GetValues is faster it can be used but for cases where GetValues is problematic or slower (e.g. because of type inference in SQLite) then the typed APIs can be used instead.

Note that we still call GetFieldValue<object> in some cases (e.g. when extracting shadow values or composite key values), but we could do similar code-gen as we do for materialization to avoid this.