I like to ride my tricycle, I like to ride my trike... (Sidecars for storing alternate property values with state entries)
Original values are additional values that can be stored alongside the state entry and updated from or merged with the actual property values when appropriate. The update pipeline also needs to store additional values for properties that are store generated such that they can be used by the update pipeline while changes are being saved. These values cannot be stored directly in the state entries because they should be discarded if the save fails--or when the transaction is rolled back once we have transactions.

These two requirements are now captured into a single general-purpose concept of the sidecar on the state entry. This means that the state manager can be responsible for store value handling such that the data store/update pipeline doesn't need to do anything special--it just propagates values back into the state entry and can likewise read values from the state entry without fear of corrupting the state entry if the transaction is rolled back.

In addition, this change moves original values functionality out of the state entry for better separation of concerns, which in turn means that it is cleaner to change the storage for original values to be something other than just an object array if we determine that doing so is beneficial in terms of memory and/or speed.