Ground and neutral are different... (Rewire services and configuration)
This change creates a separation between the services and settings for a context. This allows EF to work better with IServiceProvider instances created externally to EF which may change (due to scoping) each time used.

At the low level the steps for services are now:
- Create a ServiceCollection and configure services on it using EF and other extension methods
- Build the provider, either explicitly, or using facilities of MVC, etc.
- Expose a constrctor on the context for IServiceProvider and eitehr register the context type as transient on the DI container or create instances explicitly by passing in the IServiceProvider.

For settings, the OnConfiguring method can be used to set things like connection strings. Alternately, an EntityConfigurationBuilder can be used to create an EntityConfiguration with these kinds of settings. The IServiceProvider should be passed to EntityConfiguration.

When not using any external DI container or services, the EntityConfigurationBuilder can be used to also build the services, which will then be cached appropriately based on the services called. In this case the sugar WithServices method can be used to add providers or other types of service.

Note: Additional testing is needed for some of this code.