The nicely hoppy Inversion IPA... (Remove WithServices and allow EntityConfiguration to be injected and extended)
This change conceptually separates EntityConfiguration from the service provider and removes methods on EntityConfigurationBuilder that allow services to be specified. However, only in the case where EF is using its own internal DI things that register configuration can also inject some services transparently.

An instance of EntityConfiguration can now optionally be placed in the DI container and injected into the context. It is also possible to create different derived EntityConfigurations in order to configure different contexts, and it is possible to extent EntityConfiguration with additional app-specific properties.

Also cleaned up persistent InMemoryDataStore to be rooted in the IServiceProvider again rather than in the app domain. This works by the InMemoryDatabase being a singleton which is used by the scoped InMemoryDataStore.

Still working on some of the tests.