It's null Jim, but not as we know it... (Temporary fix to allow saving null strings)
This is for Issue #142. The fix is to specify DBNull if the value being saved is null. In general we need to create correctly typed parameters, but that change is more involved, so getting this fix in to unblock Identity.

Also added missing dependencies to allow tests to pass in Visual Studio again.