Mary-Jo-Lisa; always the same... (Allow connection strings to be pulled by name from IConfiguration)
This change allows connection strings to be stored in any format supported by IConfiguration and then retrieved using just the name or the "name=" syntax in UseSqlServer and equivalent.

This change doesn't include anything to get IConfiguration registered and it doesn't include a way to get connection strings out of app.config/web.config, although that could potentially be done with an appropriate IConfiguration source.