Playing my recorder... (Instrument SQL Server sequence concurrency test)
I have been unable to repro the concurrency failure (Issue #266) that we occasionally see on the CI machine. This change re-enables the test with some instrumentation such that some more information can be obtained when it next fails. To start with the information gathered is the sequence blocks actually returned from SQL Server. Hopefully this will confirm that SQL Server is never returning duplicates which then narrows the scope of the problem.