More Monster Munch... (Fixup FKs using navigations when entity is first tracked)
This allows a graph to be specified using navigation properties only with the state manager taking care of setting FK properties to appropriate values.

Note that FK properties that are also PK properties must still be set because the state manager requires a unique key. This is an area that needs more work.