That name no longer has any meaning for me... (Remove parameterless overloads of ForeignKey and ReferencedKey)
After discussion of this API we decided that the semantics of these overloads were ambiguous as to whether they were for just specifying the end or indicating that the FK should be created explicitly in shadow state. Also, it was not super-clear which was the principal/dependent when no property was supplied. Given that the only real value is to avoid having to be explicit about either the FK or PK in cases where the principal/dependent need to be flipped, we decided it was better to remove these overloads.