Quicker, easier, more seductive... (Change DbContext.OnModelCreating to get a ConventionModelBuilder)
The relationship fluent API patterns we have introduced cannot work without some conventions. However, we feel that there is still value in having a ModelBuilder API that is very close to the metadata model and never uses any conventions. Therefore, the plan is to move the new relationship API into the ConventionModelBuilder. Since we want this API to be the one available in OnModelCreating this means that the parameter in OnModelCreating should now be ConventionModelBuilder.

Also did the rename of the parameter per issue #441.