A CollectionFixture seems the right way to start and stop the redis server for the tests, but they are not working in Xunit at the moment so this is the next best thing. Have added a RedisXunitTestExecutor which starts the server in its constructor and stops it in the destructor (unfortunately had to use that because Dispose() wasn't always called).
Also found that you _must_ have only one ConnectionMultiplexer per set of
config options - more causes multi-threading issues which means the tests
sometimes work but more often fail. So set up a static Dictionary of
config options to ConnectionMultiplexer in RedisDatabase and only create
the ConnectionMultiplexer as needed.

The Simple Tests are now running with server starting and stopping as
expected both under VS and for k test - so re-enabling the latter.