Hiring a temp... (Allow property values to be marked a temporary)
This is a small change to allow property values in Added entities to be marked as temporary. The code reuses generalized IsModified storage since that flag is only used in Modified entities while this flag is only used in Added entities. See https://github.com/aspnet/EntityFramework/wiki/Entity-Framework-Design-Meeting-Notes---July-10,-2014#value-generation