Random sequential access...
Add SQL Server-specific model APIs for sequences and clustering

This adds SQL Server-specific fluent and core metadata APIs as discussed in the design meeting here: https://github.com/aspnet/EntityFramework/wiki/Design-Meeting-Notes---October-2,-2014

The sequence APIs are different from discussed in the meeting because they now more explicitly separate the use of sequences for value generation from the definition of sequences in the relational model. The definition of sequences now has a relational-specific fluent API and this is the only place where characteristics of the sequence can be changed. The model can then be setup to use sequences for value generation, as can each property.

This code doesn't include connecting up the new annotations to the Migrations code.