Fix break that made SQL Server keys non-clustered by default
Or: Nuclear collapse of the tri-valued cluster bomb

Issue #911
The clustered annotation is now SQL Server specific and default to null, meaning not specified. The Migrations database builder does not have the concept of not specified and was interpreting null as false. This resulted in keys being created explicitly as non-clustered. This change makes null be treated as clustered by default to revert the break, but note that the core/Migrations difference in how clustering is supported still needs to be addressed--see issue #879