fixes to include-related bugs - 925 and 926
Bug #925 - include fails for entities with composite keys
Root cause of the issue is that at one point during query generation we create join predicate expressions using Expression.And operator, but later in the pipeline we expect AndAlso operator.
Fix it to construct the initial expression using AndAlso operators.

Bug #926 - invalid sql generated when performing query with include
Root cause of the issue is that we build the comparison for join predicate incorrectly, swapping the column names.
Fix it to swap the names to the correct order.

CR: Andrew